+++ assets/data/Navbar.json
@@ -1,0 +1,24 @@
+{
+    "navbarData": [
+        {
+            "label": "Intro",
+            "icon": "home",
+            "href": "/"
+        },
+        {
+            "label": "Portfolio",
+            "icon": "th",
+            "href": "/details-page"
+        },
+        {
+            "label": "About Me",
+            "icon": "user",
+            "href": "/details-page"
+        },
+        {
+            "label": "Contact",
+            "icon": "envelope",
+            "href": "/details-page"
+        }
+    ]
+}
