+++ components/Header/index.module.scss
@@ -1,0 +1,17 @@
+.header {
+    position: fixed;
+    left: 0;
+    top: 0;
+    bottom: 0;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 20%;
+    height: 100%;
+    overflow-y: auto;
+    text-align: right;
+    padding: 2rem;
+    background: #222629 url('../../assets/images/overlay.png');
+    box-shadow: inset -0.25em 0 0.25em 0 rgba(0, 0, 0, 0.1);
+    color: #fff;
+}
