--- pages/Landing.tsx
@@ -1,9 +1,0 @@
-import styles from '../styles/Home.module.css'
-
-export default function Home() {
-  return (
-    <div className={styles.container}>
-        <h1>Landing</h1>
-    </div>
-  )
-}
