+++ styles/App.module.scss
@@ -9,6 +9,7 @@
   color: red;
 }
 .container {
+  margin-left: 20%;
   padding: 0 2rem;
 }
 
@@ -22,23 +23,6 @@
   align-items: center;
 }
 
-.header {
-  position: fixed;
-  left: 0;
-  top: 0;
-  bottom: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 20%;
-  height: 100%;
-  overflow-y: auto;
-  text-align: right;
-  padding: 2rem;
-  background: #222629 url('../assets/images/overlay.png');
-  box-shadow: inset -0.25em 0 0.25em 0 rgba(0, 0, 0, 0.1);
-  color: #fff;
-}
 
 .footer {
   display: flex;
