+++ utilities/iconMap.tsx
@@ -1,0 +1,12 @@
+// utilities/iconMap.js
+import { faHome, faTh, faUser, faEnvelope /* Add more icons here */ } from '@fortawesome/free-solid-svg-icons';
+
+const iconMap = {
+  "home": faHome,
+  "th": faTh,
+  "user": faUser,
+  "envelope": faEnvelope
+  // Add more icons here, matching the icon name to the imported icon
+};
+
+export default iconMap;
